+++ SistemaControleEmpresarial.layout
@@ -1,20 +1,20 @@
 [Editor_0]
-CursorCol=22
-CursorRow=150
-TopLine=143
+CursorCol=27
+CursorRow=17
+TopLine=305
 LeftChar=1
 [Editors]
-Order=5,2,1,3,4
-Focused=1
+Order=0,5,2,1,4,3
+Focused=4
 [Editor_1]
-CursorCol=2
-CursorRow=657
-TopLine=83
+CursorCol=17
+CursorRow=149
+TopLine=380
 LeftChar=1
 [Editor_2]
 CursorCol=4
 CursorRow=255
-TopLine=246
+TopLine=1
 LeftChar=1
 [Editor_3]
 CursorCol=2
@@ -23,11 +23,11 @@ TopLine=91
 LeftChar=1
 [Editor_4]
 CursorCol=2
-CursorRow=567
-TopLine=552
+CursorRow=340
+TopLine=328
 LeftChar=1
 [Editor_5]
-CursorCol=20
-CursorRow=81
-TopLine=118
+CursorCol=4
+CursorRow=126
+TopLine=16
 LeftChar=1
