+++ menu.c
@@ -4,39 +4,11 @@
 #include <stdlib.h>
 #include <ctype.h>
 
+#include "menu.h"
+#include "cadastro.h"
+#include "interface.h"
+
 #define TAM_TITULO_MENU 40
-
-//Definição do tipo OpcaoDoMenu
-typedef struct {
-	char textoDaOpcao[50];
-	int numeroDaOpcao;
-	int (*funcao)();
-} OpcaoDoMenu;
-
-//Definição do tipo Menu
-typedef struct {
-	char tituloDoMenu[50];
-	int numeroDeOpcoes;
-	OpcaoDoMenu opcoes[10];
-} Menu;
-
-
-//Declaração das funções de chamada das opções
-	//Opções do Menu
-		int exibirMenuPrincipal();
-			int exibirMenuCadastro();
-				int exibirMenuNovoCadastro();
-				int exibirMenuConsultarCadastro();
-			
-			int exibirMenuOrdemProducao();
-				int exibirMenuNovaOrdemProducao();
-				int exibirMenuConsultarOrdemProducao();
-	//
-
-	//Função de fim do Menu
-		int fimDoMenu();
-	//
-//
 
 //Criação dos Menus
 /* A criação dos Menus segue a seguinte estrutura:
@@ -45,22 +17,21 @@ typedef struct {
 		"tituloDoMenu", [String]
 		numeroDeOpçõesDesseMenu, [Int]
 		{ //Inicialização das opções do Menu
-			{"Texto da opção", numeroDaOpção, funçãoASerExecutada}, [String, Int, Function]
+			{"Texto da opção", numeroDaOpção, &funçãoASerExecutada}, [String, Int, Function]
 			...
 			...
-			{"Texto da opção", numeroDaOpção, funçãoASerExecutada}, [String, Int, Function]
+			{"Texto da opção", numeroDaOpção, &funçãoASerExecutada}, [String, Int, Function]
 		}
 	};
 	
 */
-
 	//menuPrincipal
 		Menu menuPrincipal = {
 			"MENU",
-			3,
+			2,
 			{
 				{"Cadastros", 1, &exibirMenuCadastro},
-				{"Ordem de Produção", 2, &exibirMenuOrdemProducao},
+//				{"Ordem de Produção", 2, &exibirMenuOrdemProducao},
 				{"Sair", 0, &fimDoMenu},
 			}
 		};
@@ -72,7 +43,7 @@ typedef struct {
 			3,
 			{
 				{"Criar novo Cadastro", 1, &exibirMenuNovoCadastro},
-				{"Consultar Cadastros", 2, &exibirMenuConsultarCadastro},
+				{"Consultar e Modificar Cadastros", 2, &exibirMenuConsultarCadastro},
 				{"Voltar ao Menu Inicial", 0, &exibirMenuPrincipal},
 			}
 		};
@@ -80,8 +51,11 @@ typedef struct {
 			//Criar novo Cadastro
 				Menu menuNovoCadastro = {
 					"NOVO CADASTRO",
-					2,
+					5,
 					{
+						"Clientes", 1, &executarNovoCadastro,
+						"Funcionários", 2, &executarNovoCadastro,
+						"Fornecedores", 3, &executarNovoCadastro,
 						"Voltar ao Menu Anterior", 0, &exibirMenuCadastro,
 						"Voltar ao Menu Inicial", 9, &exibirMenuPrincipal,
 					}
@@ -90,16 +64,33 @@ typedef struct {
 			
 			//Consultar Cadastros
 				Menu menuConsultarCadastro = {
-					"CONSULTAR CADASTROS",
-					2,
+					"CONSULTAR E MODIFICAR CADASTROS",
+					5,
 					{
+						"Clientes", 1, &executarConsultaCadastro,
+						"Funcionários", 2, & executarConsultaCadastro,
+						"Fornecedores", 3, &executarConsultaCadastro,
 						"Voltar ao Menu Anterior", 0, &exibirMenuCadastro,
 						"Voltar ao Menu Inicial", 9, &exibirMenuPrincipal,
 					}
 				};
+				
+				//Menu de modificações do cadastro
+					Menu menuModificarCadastro = {
+						"CONSULTAR E MODIFICAR CADASTROS",
+						5,
+						{
+							"Alterar Campo Específico do Cadastro", 1, &executarAlterarCadastro,
+							"Alterar Todo o Cadastro", 2, &exibirMenuCadastro,
+							"Deletar o Cadastro", 3, &exibirMenuCadastro,
+							"Consultar Outro", 0, &exibirMenuConsultarCadastro,
+							"Voltar ao Menu Principal", 9, &exibirMenuPrincipal,
+						}
+					};
 			//
 	//
 	
+/** Para uso Posterior
 	//menuOrdemProducao
 		Menu menuOrdemProducao = {
 			"ORDEM PRODUÇÃO",
@@ -133,183 +124,219 @@ typedef struct {
 				};
 			//
 	//
+**/
 //
 
-
-//Declaração de funções de uso
-	//Usada para receber o input do usuário e retornar erro caso seja digitado uma letra, ou um número que não tem no menu
-	int receberOpcaoMenu(int *pEscolhaUser, int *pErr, Menu *pMenuAtual);
+//Criação de variáveis de escopo global para manipulação em funções
+	Menu menuAtual; //Variável do tipo Menu que armazenará qual é o Menu Atual do programa.
+	Menu *pMenuAtual = &menuAtual;
 	
-	//Exibe uma mensagem de erro de acordo com o erro passado
-	void exibirErro(int *pErr);
+	int escolhaUser; //Variável que armazenará qual é o valor digitado pelo usuário nas opções do Menu.
+	int *pEscolhaUser = &escolhaUser;
 	
-	//Exibe o título do Menu Atual com caracteres # do lado direito e lado esquerdo, para ficar mais bonito
-	void exibirTituloMenuAtual(Menu *pMenuAtual);
+	int loopMenu = 1; //Variável responsável por controlar o loop do 'do while' dentro da função main, sendo o único jeito de encerrar o menu, caso seja digitado a opção 0 dentro do menu principal.
 
-	//Exibe as opções disponíveis no Menu Atual
-	void exibirOpcoesMenuAtual(Menu *pMenuAtual);
-
+	int mensagem = 0; //Variável que armazena o valor atual de Mensagem
+	int *pMensagem = &mensagem;	//Definição de ponteiro para poder passar a mensagem como parâmetro em funções
+	
+	int erro = 0; //Variável que armazena o valor atual de Erro
+	int *pErro = &erro; //Definição de ponteiro para poder passar o erro como parâmetro em funções
 //
-	
-	//Criação de variáveis de escopo global para manipulação em funções
-		Menu menuAtual; //Variável do tipo Menu que armazenará qual é o Menu Atual do programa.
-		int escolhaUser; //Variável que armazenará qual é o valor digitado pelo usuário nas opções do Menu.
-		int loopMenu = 1; //Variável responsável por controlar o loop do 'do while' dentro da função main, sendo o único jeito de encerrar o menu, caso seja digitado a opção 0 dentro do menu principal.
-	//
 
-int main() {
+void main() {
 	setlocale(LC_ALL, "Portuguese");
 	
 	menuAtual = menuPrincipal;
-	
-	//Setando valores genéricos para as variáveis importantes
-	int erro = 0;
-	
-	//Ponteiros das variáveis importantes
-	int *pErro = &erro;
-	int *pEscolhaUser = &escolhaUser;
-	Menu *pMenuAtual = &menuAtual;
-	
-	//Estrutura de repetição do Menu
+
 	do{
-		system("cls"); //Limpando a tela do prompt
+	    system("color E0"); //Definição da cor do sistema
+	    
+	    // Comente esse código para tirar a trava de tela
+		    HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE);
+			SMALL_RECT rect = {0, 0, 101, 42}; // Define o tamanho da janela
+			SetConsoleWindowInfo(hConsole, TRUE, &rect); // Define o tamanho da janela
+			SetWindowLong(GetConsoleWindow(), GWL_STYLE, GetWindowLong(GetConsoleWindow(), GWL_STYLE) & ~WS_MAXIMIZEBOX & ~WS_SIZEBOX);
+    	//Comente esse código para tirar a trava de tela
+    	
+		exibirInterfaceTitulo(pMenuAtual -> tituloDoMenu, 1);		
 		
-		exibirTituloMenuAtual(pMenuAtual);
+		exibirMensagem(pMensagem); //Exibindo eventuais mensagens, por padrão erro = 0, portanto, sem mensagens
 		
 		exibirErro(pErro); //Exibindo eventuais erros, por padrão erro = 0, portanto, sem erros
 		
-		exibirOpcoesMenuAtual(pMenuAtual);
+		exibirInterfaceOpcoes(pMenuAtual);
 		
-		receberOpcaoMenu(pEscolhaUser, pErro, pMenuAtual); //Solicita a escolha do usuário e fazer o tratamento de erros
+		receberOpcaoMenu(pEscolhaUser, pErro, pMenuAtual); //Solicita a escolha do usuário e faz o tratamento de erros
 		
 	} while (loopMenu != 0);
 	
 	printf("\n\nPrograma encerrado com sucesso!");
-	return 0;
 }
 
-//Função que exibe o título do menuAtual, foi feita separada do resto para poder ser utilizada mesmo que não exista um menu.
-void exibirTituloMenuAtual(Menu *pMenuAtual){
-	
-	//Vars
-		int i;
-		int limiteCharTitulo = (TAM_TITULO_MENU - strlen(pMenuAtual -> tituloDoMenu) - 2)/2; //limite de caracteres # = tamanho máximo do menu (definido no começo do código - o tamanho do titulo atual - 2(espaços em branco adicionados no printf(" %s ") tudo isso divido por 2, pois serão feitas duas vezes, para que fique a mesma quantidade de # do lado esquerdo e do lado direito
-	//
-	
-	//For do lado esquerdo
-	for(i = 0; i < limiteCharTitulo; i++) {
-		putchar('#');
-	}
-	
-	//Titulo do Menu
-	printf(" %s ", pMenuAtual -> tituloDoMenu);
-	
-	//For do lado direito
-	for(i = 0; i < limiteCharTitulo; i++) {
-		putchar('#');
-	}
-	
-	//Pulando linhas depois do titulo completo
-	printf("\n\n");
-}
-
-//Função que exibe as opções dentro do menuAtual
-void exibirOpcoesMenuAtual(Menu *pMenuAtual) {
-	int i; 
-	
-	for(i = 0; i < pMenuAtual -> numeroDeOpcoes; i++) {
-		printf("[%d] - %s\n", pMenuAtual -> opcoes[i].numeroDaOpcao, pMenuAtual -> opcoes[i].textoDaOpcao);
-	}
-}
-
-//Função que recebe o input do usuário e trata erros de entrada inválida ou opção inexistente
-int receberOpcaoMenu(int *pEscolhaUser, int *pErro, Menu *pMenuAtual) {
+void receberOpcaoMenu(int *pEscolhaUser, int *pErro, Menu *pMenuAtual) {
 	
 	int i;
-	
-	printf("Opção: ");
 	int valid = scanf("%d", pEscolhaUser);
 	fflush(stdin);
 	
 	if(valid == 1) {
 		if(*pEscolhaUser < 0) {
 			*pErro = 2;
-			return 0;
 		} else {
 			*pErro = 2; //Assume todos os valores retornando um erro de número inválido até que seja conferido pelo for
 			for(i = 0; i < pMenuAtual -> numeroDeOpcoes; i++) { // Laço para percorrer todas as opções do Menu e ver se existe a opção digitada
 				if(pMenuAtual -> opcoes[i].numeroDaOpcao == *pEscolhaUser) { //Caso exista alguma opção que o usuário digitou
 					*pErro = 0; //Reatribui o valor do erro
-					pMenuAtual -> opcoes[i].funcao(); //Executa da função presente no item do menu
+					pMenuAtual -> opcoes[i].funcao(pMenuAtual, pErro); //Executa da função presente no item do menu
 					break; //Encerra o for
 				}
 			}
-			return 0;
 		}
 	} else {
 		*pErro = 1;
-		return 0;
 	}
 }
 
-//Função de exibição de erros
+void exibirMensagem(int *pMensagem) {
+	char res[50];
+	
+	switch(*pMensagem) {
+		case 1:
+			strcpy(res, "Sucesso! Cadastro Concluído!");
+			break;
+		case 2:
+			strcpy(res, "Sucesso! Campo alterado!");
+			break;
+		case 3:
+			strcpy(res, "Cadastro Excluído!");
+			break;
+		default:;
+	}
+	
+	if(*pMensagem != 0) exibirInterfaceInteracao(res);
+	*pMensagem = 0;
+}
+
 void exibirErro(int *pErro) {
+	char res[50];
 	switch(*pErro) {
 		case 0:
 			break;
 		case 1:
-			printf("Digite apenas números!\n");
+			strcpy(res, "Erro: Digite apenas números!");
 			break;
 		case 2:
-			printf("Opção inválida! Digite novamente!\n");
+			strcpy(res, "Erro: Opção inválida! Digite novamente!");
 		break;
-		default:
+		case 3:
+			strcpy(res, "Erro: Não foi possível realizar sua consulta!");
 			break;
+		case 4:
+			strcpy(res, "Erro: Não existe nenhum registro para esse Número de Cadastro!");
+			break;
+		default:;
 	}
+	
+	if(*pErro != 0) exibirInterfaceInteracao(res);
 	*pErro = 0;
 }
 
 //Funções a serem executadas pelas opções dos Menus
 	//Funções de chamada do Menu
-		int exibirMenuPrincipal() {
+		void exibirMenuPrincipal() {
 			menuAtual = menuPrincipal;
 		}
 		
 			//Funções menuCadastro
-				int exibirMenuCadastro() {
-					menuAtual = menuCadastro;	
+				void exibirMenuCadastro() {
+					menuAtual = menuCadastro;
 				}
 				
-					int exibirMenuNovoCadastro() {
+					void exibirMenuNovoCadastro() {
 						menuAtual = menuNovoCadastro;
 					}
+						void executarNovoCadastro(Menu *pMenuAtual, int *pErro) {
+							novoCadastro(pEscolhaUser, pMensagem, pErro);
+						}
 					
-					int exibirMenuConsultarCadastro() {
+					void exibirMenuConsultarCadastro() {
 						menuAtual = menuConsultarCadastro;
 					}
+						void executarConsultaCadastro(Menu *pMenuAtual, int *pErro) {
+							int estado = 0;
+							int *pEstado = &estado;
+							consultaCadastro(pEscolhaUser, pMensagem, pErro, pEstado);
+							if(estado == 1) {
+								receberOpcoesAlteracaoCadastro();
+							}
+						}
+							void executarAlterarCadastro() {
+								exibirInterfaceTitulo(menuAtual.tituloDoMenu, 1);
+								alterarCadastro(pEscolhaUser, pMensagem, pErro);
+							}
+						
 			//
 			
+		/** Para uso posterior
 			//Funções menuOrdemProducao
-				int exibirMenuOrdemProducao() {
+				void exibirMenuOrdemProducao() {
 					menuAtual = menuOrdemProducao;
 				}
-					int exibirMenuNovaOrdemProducao() {
+					void exibirMenuNovaOrdemProducao() {
 						menuAtual = menuNovaOrdemProducao;
 					}
 					
-					int exibirMenuConsultarOrdemProducao() {
+					void exibirMenuConsultarOrdemProducao() {
 						menuAtual = menuConsultarOrdemProducao;
 					}
 			//
+		**/
 	//
 	
 	//Função de fim do Menu
-		int fimDoMenu() {
+		void fimDoMenu() {
 			loopMenu = 0;
-			return 0;
 		}
 	//
-	
+
+//Função de recebimento de opções na parte de Alteração de cadastro
+	void receberOpcoesAlteracaoCadastro() {
+		int escolhaUserAlterar = 1;
+		int *pEscolhaUserAlterar = &escolhaUserAlterar;
+		
+		while(escolhaUserAlterar == 1 || escolhaUserAlterar == 2 || escolhaUserAlterar == 3) {
+			menuAtual = menuModificarCadastro;
+			exibirInterfaceOpcoes(pMenuAtual);
+			exibirMensagem(pMensagem);
+			scanf("%d", &escolhaUserAlterar);
+			fflush(stdin);
+			
+			switch(escolhaUserAlterar) {
+				case 1:
+					executarAlterarCadastro();
+					break;
+				case 2:
+					executarExcluirCadastro(2, pEscolhaUserAlterar, pMensagem, pErro);
+					if(escolhaUserAlterar == 0) {
+						exibirMenuNovoCadastro();
+						executarNovoCadastro(pMenuAtual, pErro);
+						menuAtual = menuConsultarCadastro;
+						executarConsultaCadastro(pMenuAtual, pErro);
+					}
+					break;
+				case 3:
+					executarExcluirCadastro(1, pEscolhaUserAlterar, pMensagem, pErro);
+					if(escolhaUserAlterar == 0) {
+						menuAtual = menuConsultarCadastro;
+						executarConsultaCadastro(pMenuAtual, pErro);
+					}
+					break;
+				case 0:
+					exibirMenuConsultarCadastro();
+					break;
+				default:
+					exibirMenuPrincipal();
+			}
+		}
+	}
 //
-
-
