--- testesFuncoes/testesInterface/ProjetoTesteInterface.layout
@@ -1,18 +1,0 @@
-[Editor_0]
-CursorCol=2
-CursorRow=8
-TopLine=1
-LeftChar=1
-[Editor_2]
-CursorCol=29
-CursorRow=18
-TopLine=40
-LeftChar=1
-[Editors]
-Order=1
-Focused=0
-[Editor_1]
-CursorCol=7
-CursorRow=12
-TopLine=1
-LeftChar=1
