--- testesFuncoes/testesInterface/main.c
@@ -1,75 +1,0 @@
-#include <stdio.h>
-#include <locale.h>
-#include <string.h>
-#include <stdlib.h>
-#include <ctype.h>
-
-#include "interface.h"
-
-#define MAX_STRING 50
-
-#define TAM_FORMULARIO 10
-
-Formulario formularioCadastroCliente[] = {
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Num Cadastro:     | Nome:                                                                          |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Data Nasc.:                                | CPF:                                                  |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Telefone:                                  | E-mail:                                               |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Endereço:                                                                    | Número:             |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Complemento:                               | Bairro:                                               |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| CEP:                      | Cidade:                                          | Estado:             |",
-	"+----------------------------------------------------------------------------------------------------+",
-};
-
-Formulario formularioCadastroFuncionario[] = {
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Num Cadastro:     | Nome:                                                      | Status:           |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| CPF:                            |  Data de Nasc.:               | Estado Civil:                    |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Data de Admissão:                           | Salário Base: R$                                     |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Telefone:                                   | E-mail:                                              |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Endereço:                                                                    | Número:             |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Complemento:                               | Bairro:                                               |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| CEP:                      | Cidade:                                          | Estado:             |",
-	"+----------------------------------------------------------------------------------------------------+",
-};
-
-Formulario formularioCadastroFornecedor[] = {
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Num Cadastro:     | Nome Fantasia:                                                                 |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| CNPJ:                     | Telefone:                      | E-mail:                               |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Endereço:                                                                 | Número:                |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| Complemento:                               | Bairro:                                               |",
-	"+----------------------------------------------------------------------------------------------------+",
-	"| CEP:                      | Cidade:                                          | Estado:             |",
-	"+----------------------------------------------------------------------------------------------------+",
-};
-
-int main() {
-	setlocale(LC_ALL, "Portuguese");
-	
-	char stringHolder[30];
-	
-	int contadorFormulario;
-	int *pContadorFormulario = &contadorFormulario;
-	
-	printf("Digite seu nome: ");
-	fgets(stringHolder, 30, stdin);
-	fflush(stdin);
-	stringHolder[strcspn(stringHolder, "\n")] = '\0';
-	
-	return 0;
-}
