--- testesFuncoes/testesInterface/testeInterface.layout
@@ -1,18 +1,0 @@
-[Editors]
-Order=0,1,2
-Focused=0
-[Editor_0]
-CursorCol=25
-CursorRow=35
-TopLine=18
-LeftChar=1
-[Editor_1]
-CursorCol=14
-CursorRow=9
-TopLine=1
-LeftChar=1
-[Editor_2]
-CursorCol=127
-CursorRow=11
-TopLine=1
-LeftChar=1
